refactor: HTTPCode decorator
move HTTPCode decorator from utils into its own module file.
