+++ nest/core/decorators/utils.py
@@ -1,23 +1,7 @@
 import ast
 import inspect
 
-from nest.common.constants import INJECTABLE_TOKEN, STATUS_CODE_TOKEN
-
-
-def HttpCode(status_code: int):
-    """
-    Decorator that sets the HTTP status code for a route.
-
-    Args:
-        status_code (int): The HTTP status code for the response.
-    """
-
-    def decorator(func):
-        if not hasattr(func, STATUS_CODE_TOKEN):
-            setattr(func, STATUS_CODE_TOKEN, status_code)
-        return func
-
-    return decorator
+from nest.common.constants import INJECTABLE_TOKEN
 
 
 def get_instance_variables(cls):
